requirements.txt
streamlit==1.25.0
joblib==1.4.2
numpy==1.23.5
pandas==1.5.3
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
requests
python-dotenv
transformers
torch
faiss-cpu
sentence-transformers
streamlit-lottie

